discountsite/requirements.txt
asgiref==3.5.0
Django==4.0.3
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
djangorestframework==3.13.1
djangotoolbox==1.8.0
requests
bs4
psycopg2